gcp/requirements.txt
functions-framework==3.*
tensorflow==2.15.0
google-cloud-storage==2.14.0
Pillow==11.1.0
